requirements.txt
pyrebase4
flask
flask_cors
google.generativeai
openai
setuptools
dotenv
chess
